Add default reviewers information into bitbucket server logic
