+++ modules/core/src/main/scala/org/scalasteward/core/bitbucketserver/http4s/Http4sBitbucketServerApiAlg.scala
@@ -16,12 +16,13 @@
 
 package org.scalasteward.core.bitbucketserver.http4s
 
-import cats.data.NonEmptyList
 import cats.effect.Sync
 import cats.implicits._
 import io.circe.generic.semiauto.{deriveDecoder, deriveEncoder}
 import io.circe.{Decoder, Encoder}
 import org.http4s.{Request, Uri}
+import org.scalasteward.core.bitbucketserver.http4s.Json.{Reviewer, User}
+import org.scalasteward.core.git.Branch
 import org.scalasteward.core.git.{Branch, Sha1}
 import org.scalasteward.core.util.HttpJsonClient
 import org.scalasteward.core.vcs.VCSApiAlg
@@ -33,6 +34,7 @@ import org.scalasteward.core.vcs.data._
   */
 class Http4sBitbucketServerApiAlg[F[_]: Sync](
     bitbucketApiHost: Uri,
+    user: AuthenticatedUser,
     modify: Repo => Request[F] => F[Request[F]]
 )(implicit client: HttpJsonClient[F])
     extends VCSApiAlg[F] {
@@ -46,8 +48,9 @@ class Http4sBitbucketServerApiAlg[F[_]: Sync](
     val toRef =
       Json.Ref("refs/heads/" + data.base.name, Json.Repository(repo.repo, Json.Project(repo.owner)))
 
-    val req =
-      Json.NewPR(
+    for {
+      reviewers <- getDefaultReviewers(repo)
+      req = Json.NewPR(
         title = data.title,
         description = data.body,
         state = Open,
@@ -56,13 +59,22 @@ class Http4sBitbucketServerApiAlg[F[_]: Sync](
         fromRef = fromRef,
         toRef = toRef,
         locked = false,
-        reviewers = List.empty
+        reviewers = reviewers
       )
-
-    client
-      .postWithBody[Json.PR, Json.NewPR](url.pullRequests(repo), req, modify(repo))
-      .map(pr => PullRequestOut(pr.links("self").head.href, pr.state, pr.title))
+      pr <- client.postWithBody[Json.PR, Json.NewPR](url.pullRequests(repo), req, modify(repo))
+    } yield PullRequestOut(pr.links("self").head.href, pr.state, pr.title)
   }
+
+  def getDefaultReviewers(repo: Repo): F[List[Reviewer]] =
+    client
+      .get[List[Json.Condition]](url.reviewers(repo), modify(repo))
+      .map(conditions =>
+        conditions
+          .flatMap(condition =>
+            condition.reviewers
+              .map(reviewer => Reviewer(User(reviewer.name)))
+          )
+      )
 
   override def getBranch(repo: Repo, branch: Branch): F[BranchOut] =
     client
@@ -84,74 +96,26 @@ class Http4sBitbucketServerApiAlg[F[_]: Sync](
 
   def ni(name: String): Nothing = throw new NotImplementedError(name)
 
-  object Json {
-    case class Page[A](values: List[A])
-
-    case class Repo(name: String, forkable: Boolean, project: Project, links: Links)
-
-    case class Project(key: String)
-
-    type Links = Map[String, NonEmptyList[Link]]
-
-    case class Link(href: Uri, name: Option[String])
-
-    case class PR(title: String, state: PullRequestState, links: Links)
-
-    case class NewPR(
-        title: String,
-        description: String,
-        state: PullRequestState,
-        open: Boolean,
-        closed: Boolean,
-        fromRef: Ref,
-        toRef: Ref,
-        locked: Boolean,
-        reviewers: List[Reviewer]
-    )
-
-    case class Ref(id: String, repository: Repository)
-
-    case class Repository(slug: String, project: Project)
-
-    case class Reviewer(user: User)
-
-    case class User(name: String)
-
-    case class Branches(values: NonEmptyList[Branch])
-
-    case class Branch(id: String, latestCommit: Sha1)
-
-    implicit def pageDecode[A: Decoder]: Decoder[Page[A]] = deriveDecoder
-    implicit val repoDecode: Decoder[Repo] = deriveDecoder
-    implicit val projectDecode: Decoder[Project] = deriveDecoder
-    implicit val linkDecoder: Decoder[Link] = deriveDecoder
-    implicit val uriDecoder: Decoder[Uri] = Decoder.decodeString.map(Uri.unsafeFromString)
-    implicit val prDecoder: Decoder[PR] = deriveDecoder
-    implicit val branchDecoder: Decoder[Branch] = deriveDecoder
-    implicit val branchesDecoder: Decoder[Branches] = deriveDecoder
-
-    implicit val encodeNewPR: Encoder[NewPR] = deriveEncoder
-    implicit val encodeRef: Encoder[Ref] = deriveEncoder
-    implicit val encodeRepository: Encoder[Repository] = deriveEncoder
-    implicit val encodeProject: Encoder[Project] = deriveEncoder
-    implicit val encodeReviewer: Encoder[Reviewer] = deriveEncoder
-    implicit val encodeUser: Encoder[User] = deriveEncoder
-  }
 }
 
 final class StashUrls(base: Uri) {
   val api: Uri = base / "rest" / "api" / "1.0"
+  val reviewerApi: Uri = base / "rest" / "default-reviewers" / "1.0"
 
   def repo(repo: Repo): Uri =
     api / "projects" / repo.owner / "repos" / repo.repo
 
   def pullRequests(r: Repo): Uri = repo(r) / "pull-requests"
 
+  def reviewers(repo: Repo): Uri =
+    reviewerApi / "projects" / repo.owner / "repos" / repo.repo / "conditions"
+
   def listPullRequests(r: Repo, head: String): Uri =
     pullRequests(r)
       .withQueryParam("at", head)
       .withQueryParam("limit", "1000")
       .withQueryParam("direction", "outgoing")
+
 
   def branches(r: Repo): Uri = repo(r) / "branches"
